Major refactor into multiple files and begin improving types
